+++ src/fastapi_augment/config/__init__.py
@@ -1,12 +1,12 @@
 """
-@Author     : zarkhan
-@CreateDate : 2026/9/6
-@Description: 配置管理模块
+@Author         : zarkhan
+@CreateDate     : 2026/9/6
+@Description    : 配置管理模块
 """
+from .base_settings import AugmentBaseSettings
 from .database_settings import DatabaseSettings
-from .settings import EnvSettings
 
 __all__ = [
-    'DatabaseSettings',
-    'EnvSettings'
+    'AugmentBaseSettings',
+    'DatabaseSettings'
 ]